% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.r
\name{buildglmmTMB}
\alias{buildglmmTMB}
\title{Use \code{buildmer} to perform stepwise elimination on \code{glmmTMB} models}
\usage{
buildglmmTMB(
  formula,
  data = NULL,
  family = gaussian(),
  buildmerControl = buildmerControl()
)
}
\arguments{
\item{formula}{See the general documentation under \code{\link{buildmer-package}}}

\item{data}{See the general documentation under \code{\link{buildmer-package}}}

\item{family}{See the general documentation under \code{\link{buildmer-package}}}

\item{buildmerControl}{Control arguments for buildmer --- see the general documentation under \code{\link{buildmerControl}}}
}
\description{
Use \code{buildmer} to perform stepwise elimination on \code{glmmTMB} models
}
\examples{
library(buildmer)
if (requireNamespace('glmmTMB')) {
	model <- buildglmmTMB(Reaction ~ Days + (Days|Subject),data=lme4::sleepstudy)
}
}
\seealso{
\code{\link{buildmer-package}}
}
