% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\docType{data}
\name{migrant}
\alias{migrant}
\title{A very small dataset from a pilot study on sound change.}
\format{
A standard data frame.
}
\usage{
data(migrant)
}
\description{
A very small dataset from a pilot study on sound change.
}
\keyword{datasets}
